requirements.txt
aiogram==3.5.0
aiosqlite==0.20.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
